requirements.txt
# Core Django - Compatible versions
Django==4.2.15
djangorestframework==3.14.0

# Database - Stable version
psycopg2-binary==2.9.7

# Authentication & Authorization - Compatible version
django-allauth==0.57.0

# Caching & Sessions - Stable versions
redis==4.6.0
django-redis==5.3.0

# Task Queue - Compatible versions
celery==5.3.1
django-celery-beat==2.5.0
django-celery-results==2.5.1

# WebSocket Support - Stable versions
channels==4.0.0
channels-redis==4.1.0

# HTTP & API - Stable versions
requests==2.31.0
urllib3==2.0.4

# Environment & Configuration
python-decouple==3.8
python-dotenv==1.0.0

# Security - Compatible versions
django-axes==6.1.1
django-cors-headers==4.2.0
django-csp==3.7
django-ratelimit==4.0.0

# File Handling - Stable version
Pillow==10.0.0

# Production Server - Stable versions
gunicorn==21.2.0
gevent==23.7.0

# Monitoring & Logging - Compatible versions
sentry-sdk==1.32.0
django-health-check==3.17.0

# Utilities - Stable versions
pytz==2023.3
python-dateutil==2.8.2
pandas==2.0.3
openpyxl==3.1.2
reportlab==4.0.4

# Notifications - Compatible version
django-notifications-hq==1.7.0

# Static Files - Stable version
whitenoise==6.5.0

# Forms & Widgets - Compatible versions
django-crispy-forms==2.0
crispy-bootstrap5==0.7

# API Documentation - Compatible version
drf-spectacular==0.26.4

# Google APIs - Stable versions
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.97.0

# Email - Compatible version
django-ses==3.5.0

# Backup & Migration - Compatible version
django-dbbackup==4.0.2

# Production optimizations - Compatible versions
django-extensions==3.2.1
django-compressor==4.4